Switch to "chrono" for timing

+++ fahbench/Updater.h
@@ -6,7 +6,7 @@
 
 class Updater {
 public:
-    virtual void progress(int step, int total_step, double score) = 0;
+    virtual void progress(int step, int total_step, float score) = 0;
     virtual void message(std::string) = 0;
     virtual void message(boost::format) = 0;
 };
